Aggiunti controlli di ruoli. Sostituito Users con Assigners e Assignees. Modifiche a ProductBatches, ProductBatchPartitions, ShockReports (controllori, tabella e template) e allo stile
Controllers:
- ProductBatchesController: aggiunto controllo di ruoli in index, view, add, edit e delete. Introdotto il collegamento a Assignees e Assigners nelle query al posto di Users.
- ProductBatchesPartitionsController: Aggiunto parametro batchId ad add. Introdotto il collegamento a Assignees e Assigners nelle query al posto di Users. Settati i campi nulli e aggiunto il controllo sul parametro batchId e sull'id dell'utente loggato nel metodo add.
- ProductsController: modificato find('list') in find('all') nel metodo add.
-ShockReportsController: rimosse le variabili shockReportsUnprocessed e shockReportProcessed e le loro query.

Table:
-ProductBatchesTable: modificata associazione belongsTo per Assigner e Assignee (classeName Users).

Templates:
-ProductBatchPartitions:
--add: rimossi gli input che vengono settati in automatico dal controller e alcuni refusi del copia-incolla dai template di ProductBatch
--index: applicati i formati valuta e data ad alcuni input. Aggiunta delle classi bottoni di bootstrap ai link.
-ProductBatches:
--index: gestione degli array completedProductBatches e openProductBatches (dopo la loro rimozione dal controller). Suddivisione dei product batches in completed, open e all. Gestione dinamica delle info presenti nella tabella (assignee e assigner) in base al ruolo dell'utente. Formattazione di valute e di date. Aggiunta class btn-sm ai bottoni.
--view: modifica di User in Assignee e calcolo della quantità effettivamente venduta e del prezzo medio. Formattazione di valute e date. Aggiunta pulsante "aggiungi partizione".
-Products:
--add: aggiunta select per product area
-ShockReports:
--index_general_director: gestione degli array shockReportsUnprocessed e shockReportsProcessed (dopo la loro rimozione dal controller).
--index_seller: Come per index_general_director. Correzione della condizione per mostrare il botton per il delete.
--index_division_director: come per index_seller. Aggiunta class btn-sm ai bottoni.
-Users:
--index_seller: aggiunta classe btn-sm ai bottoni.
--index_division_director: come per index_seller
--index_general_manager: come per index_seller

Style (custom-style.scss): aggiunte variabili per i colori, modificato z-index per l'header e per la sidebar, aggiunto min-width alla classe actions.

+++ src/Controller/ProductBatchesController.php
@@ -20,11 +20,38 @@ class ProductBatchesController extends AppController
      */
     public function index()
     {
-        $this->paginate = [
-            'contain' => ['Products', 'Users']
-        ];
-        $productBatches = $this->paginate($this->ProductBatches);
-
+        $loggedUser = $this->getRequest()->getSession()->read("Auth.User");
+        $loggedUserRoleId = $loggedUser["role_id"];
+        $loggedUserRoleLetter = substr($loggedUserRoleId, 0, 1);
+        switch($loggedUserRoleLetter){
+            case "S":
+                $this->Flash->error(__('You are not authorized to access the requested content.'));
+                return $this->redirect(['controller'=>'ProductBatchPartitions', 'action' => 'index']);
+                break;
+            case "D":
+                //mostra i product batches assegnati all'utente
+                $productBatches = $this->ProductBatches->find('all', [
+                    'contain' => ['Products', 'Assignees', "Assigners"],
+                    'conditions' => [
+                        'ProductBatches.assignee_id' => $loggedUser["id"]
+                    ]
+                ]);
+                break;
+            case "G":
+                //mostra i product batches assegnati dall'utente
+                $productBatches = $this->ProductBatches->find('all', [
+                    'contain' => ['Products', 'Assignees', "Assigners"],
+                    'conditions' => [
+                        'ProductBatches.assigner_id' => $loggedUser["id"]
+                    ]
+                ]);
+                break;
+            default:
+                $this->Flash->error(__('You are not authorized to access the requested content.'));
+                return $this->redirect(['controller'=>'Users', 'action' => 'index']);
+        }
+        
+        $productBatches = $this->paginate($productBatches);
         $this->set(compact('productBatches'));
     }
 
@@ -36,9 +63,17 @@ class ProductBatchesController extends AppController
      * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
      */
     public function view($id = null)
-    {
+    {   
+        $loggedUser = $this->getRequest()->getSession()->read("Auth.User");
+        $loggedUserRoleId = $loggedUser["role_id"];
+        $loggedUserDept = substr($loggedUser["role_id"], 1, 1);
+        $loggedUserRoleLetter = substr($loggedUserRoleId, 0, 1);
+        if($loggedUserRoleLetter == 'S'){
+            $this->Flash->error(__('You are not authorized to access the requested content.'));
+            return $this->redirect(['controller'=>'Users', 'action' => 'index']);
+        }
         $productBatch = $this->ProductBatches->get($id, [
-            'contain' => ['Products', 'Users', 'ProductBatchPartitions']
+            'contain' => ['Products', 'Assigners', 'Assignees', 'ProductBatchPartitions']
         ]);
 
         $this->set('productBatch', $productBatch);
@@ -54,6 +89,11 @@ class ProductBatchesController extends AppController
         $loggedUser = $this->getRequest()->getSession()->read("Auth.User");
         $loggedUserRoleId = $loggedUser["role_id"];
         $loggedUserDept = substr($loggedUser["role_id"], 1, 1);
+        $loggedUserRoleLetter = substr($loggedUserRoleId, 0, 1);
+        if($loggedUserRoleLetter == 'S' || $loggedUserRoleLetter == 'D'){
+            $this->Flash->error(__('You are not authorized to access the requested content.'));
+            return $this->redirect(['controller'=>'Users', 'action' => 'index']);
+        }
         $productBatch = $this->ProductBatches->newEntity();
         if ($this->request->is('post')) {
             $productBatch = $this->ProductBatches->patchEntity($productBatch, $this->request->getData());
@@ -87,6 +127,14 @@ class ProductBatchesController extends AppController
      */
     public function edit($id = null)
     {
+        $loggedUser = $this->getRequest()->getSession()->read("Auth.User");
+        $loggedUserRoleId = $loggedUser["role_id"];
+        $loggedUserDept = substr($loggedUser["role_id"], 1, 1);
+        $loggedUserRoleLetter = substr($loggedUserRoleId, 0, 1);
+        if($loggedUserRoleLetter == 'S'){
+            $this->Flash->error(__('You are not authorized to access the requested content.'));
+            return $this->redirect(['controller'=>'Users', 'action' => 'index']);
+        }
         $productBatch = $this->ProductBatches->get($id, [
             'contain' => []
         ]);
@@ -113,6 +161,14 @@ class ProductBatchesController extends AppController
      */
     public function delete($id = null)
     {
+        $loggedUser = $this->getRequest()->getSession()->read("Auth.User");
+        $loggedUserRoleId = $loggedUser["role_id"];
+        $loggedUserDept = substr($loggedUser["role_id"], 1, 1);
+        $loggedUserRoleLetter = substr($loggedUserRoleId, 0, 1);
+        if($loggedUserRoleLetter == 'S' || $loggedUserRoleLetter == 'D'){
+            $this->Flash->error(__('You are not authorized to access the requested content.'));
+            return $this->redirect(['controller'=>'Users', 'action' => 'index']);
+        }
         $this->request->allowMethod(['post', 'delete']);
         $productBatch = $this->ProductBatches->get($id);
         if ($this->ProductBatches->delete($productBatch)) {
